+++ Source/StringExtensions.swift
@@ -10,6 +10,29 @@ import Foundation
 
 public extension String {
 
+    /// Find the string between two string.
+    ///
+    /// - Parameters:
+    ///   - left: The left bookend
+    ///   - right: The right bookend
+    /// - Returns: Return the String if found else `nil`
+    func between(_ left: String, _ right: String) -> String? {
+        guard let leftRange = range(of: left),
+            let rightRange = range(of: right, options: .backwards),
+            left != right && leftRange.upperBound != rightRange.lowerBound else {
+                return nil
+        }
+        return self[leftRange.upperBound...index(before: rightRange.lowerBound)]
+    }
+
+    /// Count number of occurences for a substring
+    ///
+    /// - Parameter str: The string to count
+    /// - Returns: Number of occurences
+    func count(_ str: String) -> Int {
+        return components(separatedBy: str).count - 1
+    }
+
     /// Return a date from current string.
     ///
     /// - parameter format: The date format
@@ -21,6 +44,43 @@ public extension String {
         df.dateFormat = format
         df.locale = locale
         return df.date(from: self)
+    }
+
+    /// Check whether the string is an email or not.
+    var isEmail: Bool {
+        let regex = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}"
+        let predicate = NSPredicate(format:"SELF MATCHES %@", regex)
+        return predicate.evaluate(with: self)
+    }
+
+    /// Check whether the string contains one or more letters.
+    public var hasLetters: Bool {
+        return rangeOfCharacter(from: .letters, options: .literal) != nil
+    }
+
+    /// Check whether the string contains one or more numbers.
+    public var hasNumbers: Bool {
+        return rangeOfCharacter(from: .decimalDigits, options: .literal) != nil
+    }
+
+    /// Check whether the string contains only letters.
+    var isAlpha: Bool {
+        for c in characters {
+            if !(c >= "a" && c <= "z") && !(c >= "A" && c <= "Z") {
+                return false
+            }
+        }
+        return true
+    }
+
+    /// Check if string contains at least one letter and one number
+    var isAlphaNumeric: Bool {
+        return components(separatedBy: .alphanumerics).joined(separator: "").characters.isEmpty
+    }
+
+    /// Return the initials of the String
+    var initials: String {
+        return self.components(separatedBy: " ").reduce("") { $0 + $1[0...0] }
     }
 
     /// Decoded Base 64 String if applicable
@@ -54,4 +114,14 @@ public extension String {
         return self.addingPercentEncoding(withAllowedCharacters: .urlHostAllowed) ?? self
     }
 
+}
+
+extension String {
+    subscript (r: CountableClosedRange<Int>) -> String {
+        get {
+            let startIndex = self.index(self.startIndex, offsetBy: r.lowerBound)
+            let endIndex = self.index(startIndex, offsetBy: r.upperBound - r.lowerBound)
+            return self[startIndex...endIndex]
+        }
+    }
 }